Breaking change: Compiler embedded in sql-macro, test-section added
A different flavour of the old compiler will be embedded in each method
for the sql-macro. This currently works only for select/query.

The backend function (execute ...) is now functional, so databases
can be queried. This would run:
(execute (sql [col1 col2] table1 (or (> col1 col2)
                                 (and (> col1 5) (< col2 5)))))

In ClojureQL.clj I've added a small section about testing, with data
to replicate our test-db, good for reporting bugs.